+++ airbyte-cdk/python/airbyte_cdk/sources/streams/concurrent/thread_based_concurrent_stream.py
@@ -69,15 +69,16 @@ class ThreadBasedConcurrentStream(AbstractStream):
         Algorithm:
         1. Submit a future to generate the stream's partition to process.
           - This has to be done asynchronously because we sometimes need to submit requests to the API to generate all partitions (eg for substreams).
-          - The future will add the partitions to process on a work queue
+          - The future will add the partitions to process on a work queue.
         2. Continuously poll work from the work queue until all partitions are generated and processed
+          - If the next work item is an Exception, stop the threadpool and raise it.
           - If the next work item is a partition, submit a future to process it.
-            - The future will add the records to emit on the work queue
-            - Add the partitions to the partitions_to_done dict so we know it needs to complete for the sync to succeed
-          - If the next work item is a record, yield the record
-          - If the next work item is PARTITIONS_GENERATED_SENTINEL, all the partitions were generated
-          - If the next work item is a PartitionCompleteSentinel, a partition is done processing
-            - Update the value in partitions_to_done to True so we know the partition is completed
+            - The future will add the records to emit on the work queue.
+            - Add the partitions to the partitions_to_done dict so we know it needs to complete for the sync to succeed.
+          - If the next work item is a record, yield the record.
+          - If the next work item is PARTITIONS_GENERATED_SENTINEL, all the partitions were generated.
+          - If the next work item is a PartitionCompleteSentinel, a partition is done processing.
+            - Update the value in partitions_to_done to True so we know the partition is completed.
         """
         self._logger.debug(f"Processing stream slices for {self.name} (sync_mode: full_refresh)")
         futures: List[Future[Any]] = []
@@ -93,26 +94,32 @@ class ThreadBasedConcurrentStream(AbstractStream):
         partitions_to_done: Dict[Partition, bool] = {}
 
         finished_partitions = False
-        while record_or_partition := queue.get(block=True, timeout=self._timeout_seconds):
-            if record_or_partition == PARTITIONS_GENERATED_SENTINEL:
+        while record_or_partition_or_exception := queue.get(block=True, timeout=self._timeout_seconds):
+            if isinstance(record_or_partition_or_exception, Exception):
+                # An exception was raised while processing the stream
+                # Stop the threadpool and raise it
+                self._stop_and_raise_exception(record_or_partition_or_exception)
+            elif record_or_partition_or_exception == PARTITIONS_GENERATED_SENTINEL:
                 # All partitions were generated
                 finished_partitions = True
-            elif isinstance(record_or_partition, PartitionCompleteSentinel):
+            elif isinstance(record_or_partition_or_exception, PartitionCompleteSentinel):
                 # All records for a partition were generated
-                if record_or_partition.partition not in partitions_to_done:
+                if record_or_partition_or_exception.partition not in partitions_to_done:
                     raise RuntimeError(
-                        f"Received sentinel for partition {record_or_partition.partition} that was not in partitions. This is indicative of a bug in the CDK. Please contact support.partitions:\n{partitions_to_done}"
+                        f"Received sentinel for partition {record_or_partition_or_exception.partition} that was not in partitions. This is indicative of a bug in the CDK. Please contact support.partitions:\n{partitions_to_done}"
                     )
-                partitions_to_done[record_or_partition.partition] = True
-            elif isinstance(record_or_partition, Record):
+                partitions_to_done[record_or_partition_or_exception.partition] = True
+            elif isinstance(record_or_partition_or_exception, Record):
                 # Emit records
-                yield record_or_partition
-            elif isinstance(record_or_partition, Partition):
+                yield record_or_partition_or_exception
+            elif isinstance(record_or_partition_or_exception, Partition):
                 # A new partition was generated and must be processed
-                partitions_to_done[record_or_partition] = False
+                partitions_to_done[record_or_partition_or_exception] = False
                 if self._slice_logger.should_log_slice_message(self._logger):
-                    self._message_repository.emit_message(self._slice_logger.create_slice_log_message(record_or_partition.to_slice()))
-                self._submit_task(futures, partition_reader.process_partition, record_or_partition)
+                    self._message_repository.emit_message(
+                        self._slice_logger.create_slice_log_message(record_or_partition_or_exception.to_slice())
+                    )
+                self._submit_task(futures, partition_reader.process_partition, record_or_partition_or_exception)
             if finished_partitions and all(partitions_to_done.values()):
                 # All partitions were generated and process. We're done here
                 break
@@ -135,10 +142,17 @@ class ThreadBasedConcurrentStream(AbstractStream):
     def _check_for_errors(self, futures: List[Future[Any]]) -> None:
         exceptions_from_futures = [f for f in [future.exception() for future in futures] if f is not None]
         if exceptions_from_futures:
-            raise RuntimeError(f"Failed reading from stream {self.name} with errors: {exceptions_from_futures}")
-        futures_not_done = [f for f in futures if not f.done()]
-        if futures_not_done:
-            raise RuntimeError(f"Failed reading from stream {self.name} with futures not done: {futures_not_done}")
+            exception = RuntimeError(f"Failed reading from stream {self.name} with errors: {exceptions_from_futures}")
+            self._stop_and_raise_exception(exception)
+        else:
+            futures_not_done = [f for f in futures if not f.done()]
+            if futures_not_done:
+                exception = RuntimeError(f"Failed reading from stream {self.name} with futures not done: {futures_not_done}")
+                self._stop_and_raise_exception(exception)
+
+    def _stop_and_raise_exception(self, exception: BaseException) -> None:
+        self._threadpool.shutdown(wait=False, cancel_futures=True)
+        raise exception
 
     @property
     def name(self) -> str: